requirements.txt
asgiref==3.8.1
chardet==5.2.0
dateparser==1.2.0
Django==5.0.4
django-csvimport==3.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
tzlocal==5.2
psycopg2-binary
djangorestframework
drf-yasg
